pyproject.toml
[tool.poetry]
name = "ramochka"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
telegram = "^0.0.1"
python-telegram-bot = "^21.10"
aiohttp = "^3.11.11"
flask = "^3.1.0"
pillow = "^11.1.0"
watchdog = "^6.0.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
colorlog = "^6.9.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
